requirements.txt
asgiref==3.2.5
Django==3.0.4
django-uuslug==1.2.0
mysqlclient==1.4.6
python-slugify==4.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
